refactor: rename dim.py to axis.py
